Fix and restructure unit tests
