+++ SkynetServer.sln
@@ -44,6 +44,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docker", "docker", "{8C2BDE
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkynetServer.Tests", "tests\SkynetServer.Tests\SkynetServer.Tests.csproj", "{159E952E-2E4A-45C3-936F-576FB5E1EF52}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkynetServer.Database.Tests", "tests\SkynetServer.Database.Tests\SkynetServer.Database.Tests.csproj", "{FEEE3886-3C54-4F2C-B695-55C9F850EB9E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -74,6 +76,10 @@ Global
 		{159E952E-2E4A-45C3-936F-576FB5E1EF52}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{159E952E-2E4A-45C3-936F-576FB5E1EF52}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{159E952E-2E4A-45C3-936F-576FB5E1EF52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FEEE3886-3C54-4F2C-B695-55C9F850EB9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FEEE3886-3C54-4F2C-B695-55C9F850EB9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FEEE3886-3C54-4F2C-B695-55C9F850EB9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FEEE3886-3C54-4F2C-B695-55C9F850EB9E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -85,6 +91,7 @@ Global
 		{179DB560-A20E-4A17-AC87-60D2DE11FE21} = {AF56B7E7-849C-43B1-92B5-6753710BB539}
 		{8C5F66CA-EEE8-4223-AE44-601B076CC167} = {3252B458-BDDD-4BD8-9277-BD253264D4BC}
 		{159E952E-2E4A-45C3-936F-576FB5E1EF52} = {3252B458-BDDD-4BD8-9277-BD253264D4BC}
+		{FEEE3886-3C54-4F2C-B695-55C9F850EB9E} = {3252B458-BDDD-4BD8-9277-BD253264D4BC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CCD70523-67AD-4A9B-BAF1-4BDA67255B27}
